SmartFood-pubsub/recomendador/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.4.1+cpu
pykeen==1.10.2
class-resolver==0.4.3    
pandas==2.2.2
google-cloud-storage==2.16.0
google-cloud-pubsub==2.21.0
functions-framework==3.5.0
